+++ inst/shinyapp/global.R
@@ -6,7 +6,6 @@ suppressMessages({
   library(ggplot2)
   library(ggbeeswarm)
   library(ggpmisc)
-  #library(ggpubr)
   library(ggrepel)
   library(ggquickeda)
   library(Hmisc)